fix(types): fix circular deps
- Break type-only circular dependencies by introducing leaf type modules
  and co-locating option types with their schema implementations.
- WidgetDecoration: extract `Meta` to leaf module `meta.ts` (parameterized
  by descriptor type to avoid the type-only cycle); restore `satisfies Meta`
  in `actions.ts` and `WidgetDescriptor#applyTo`. Public `Meta` in
  `types.ts` is exposed as `Meta = MetaGeneric<WidgetDescriptor>`.
- core/index.ts: keep explicit re-export of `Extension*` types from
  `./ExtensionBuilder` for an explicit public API surface.
- YfmNoteSpecs/DeflistSpecs: move schema-level option type into
  `schema.ts` as `*SchemaOptions`; keep public `*SpecsOptions` in
  `index.ts` as an alias of the schema options (matches the existing
  Tabs/Table/Cut/Checkbox convention).
- Configure check-circular-deps to surface type-only cycles
  (no longer transformed away).

Made-with: Cursor
