+++ packages/editor/src/bundle/config/dynamicModifiers.test.ts
@@ -1,7 +1,7 @@
 import {ExtensionsManager} from '../../core';
 import {SchemaDynamicModifier} from '../../core/SchemaDynamicModifier';
 import {MarkdownParserDynamicModifier} from '../../core/markdown/MarkdownParser';
-import {MarkdownSerializerDynamicModifier} from '../../core/markdown/MarkdownSerializerDynamicModifier';
+import {MarkdownSerializerDynamicModifier} from '../../core/markdown/MarkdownSerializer';
 import {convertDynamicModifiersConfigs} from '../../core/utils/dynamicModifiers';
 import {Logger2} from '../../logger';
 import {FullSpecsPreset} from '../../presets/full-specs';
