+++ packages/editor/src/extensions/additional/GPT/hooks/usePresetList.ts
@@ -3,9 +3,8 @@ import {useMemo, useRef} from 'react';
 import type {DropdownMenuItem} from '@gravity-ui/uikit';
 
 import type {CommonAnswer} from '../ErrorScreen/types';
-import type {GptDialogProps} from '../GptDialog/GptDialog';
-import type {PresetListProps} from '../PresetList/PresetList';
 import {cnGptDialogPresetList} from '../PresetList/classname';
+import type {GptDialogProps, PresetListProps} from '../types';
 
 import {useOverflowingHorizontalItems} from './useOverflowingHorizontalItems';
