+++ packages/editor/src/extensions/behavior/WidgetDecoration/actions.ts
@@ -1,12 +1,7 @@
 import type {Transaction} from 'prosemirror-state';
 
+import type {Meta} from './meta';
 import {widgetDecorationPluginKey} from './plugin-key';
-import type {Meta} from './types';
 
-export const removeDecoration = (tr: Transaction, id: string) => {
-    const meta: Meta = {
-        type: 'remove',
-        id,
-    };
-    return tr.setMeta(widgetDecorationPluginKey, meta);
-};
+export const removeDecoration = (tr: Transaction, id: string) =>
+    tr.setMeta(widgetDecorationPluginKey, {type: 'remove', id} satisfies Meta);
