+++ packages/editor/src/extensions/behavior/WidgetDecoration/meta.ts
@@ -1,0 +1,3 @@
+export type Meta<TDescriptor = unknown> =
+    | {type: 'add'; descriptor: TDescriptor}
+    | {type: 'remove'; id: string};
