+++ packages/editor/src/extensions/markdown/Deflist/DeflistSpecs/index.ts
@@ -1,5 +1,4 @@
 import deflistPlugin from '@diplodoc/transform/lib/plugins/deflist.js';
-import type {NodeSpec} from 'prosemirror-model';
 
 import type {ExtensionAuto} from '../../../../core';
 import {nodeTypeFactory} from '../../../../utils/schema';
@@ -7,6 +6,7 @@ import {nodeTypeFactory} from '../../../../utils/schema';
 import {DeflistNode} from './const';
 import {parserTokens} from './parser';
 import {getSchemaSpecs} from './schema';
+import type {DeflistSchemaOptions} from './schema';
 import {serializerTokens} from './serializer';
 
 export {DeflistNode} from './const';
@@ -14,16 +14,9 @@ export const defListType = nodeTypeFactory(DeflistNode.List);
 export const defTermType = nodeTypeFactory(DeflistNode.Term);
 export const defDescType = nodeTypeFactory(DeflistNode.Desc);
 
-export type DeflistSpecsOptions = {
-    /**
-     * @deprecated use placeholder option in BehaviorPreset instead.
-     */
-    deflistTermPlaceholder?: NonNullable<NodeSpec['placeholder']>['content'];
-    /**
-     * @deprecated use placeholder option in BehaviorPreset instead.
-     */
-    deflistDescPlaceholder?: NonNullable<NodeSpec['placeholder']>['content'];
-};
+export type {DeflistSchemaOptions} from './schema';
+
+export type DeflistSpecsOptions = DeflistSchemaOptions & {};
 
 export const DeflistSpecs: ExtensionAuto<DeflistSpecsOptions> = (builder, opts) => {
     const schemaSpecs = getSchemaSpecs(opts, builder.context.get('placeholder'));
