+++ packages/editor/src/extensions/markdown/Image/imageUrlPaste/index.ts
@@ -2,8 +2,8 @@ import {Plugin} from 'prosemirror-state';
 
 import {InputState} from 'src/utils/input-state';
 
-import type {ParseInsertedUrlAsImage} from '../../../../bundle';
 import type {ExtensionAuto} from '../../../../core';
+import type {ParseInsertedUrlAsImage} from '../../../../utils/upload';
 import {DataTransferType} from '../../../behavior/Clipboard/utils';
 import {imageType} from '../ImageSpecs';
