+++ packages/editor/src/extensions/yfm/Video/index.ts
@@ -1,16 +1,12 @@
 import type {Action, ExtensionAuto} from '../../../core';
 
 import {VideoSpecs, type VideoSpecsOptions} from './VideoSpecs';
-import type {VideoService} from './VideoSpecs/md-video';
 import {addVideo} from './actions';
+import type {VideoActionAttrs} from './actions';
 import {vAction} from './const';
 
 export {videoNodeName, videoType} from './VideoSpecs';
-
-export type VideoActionAttrs = {
-    service: VideoService;
-    url: string;
-};
+export type {VideoActionAttrs} from './actions';
 
 export type VideoOptions = VideoSpecsOptions & {};
