+++ packages/editor/src/extensions/yfm/YfmNote/YfmNoteSpecs/index.ts
@@ -1,6 +1,5 @@
 import {log} from '@diplodoc/transform/lib/log.js';
 import yfmPlugin from '@diplodoc/transform/lib/plugins/notes/index.js';
-import type {NodeSpec} from 'prosemirror-model';
 
 import {getConfig} from 'src/configure';
 import type {ExtensionAuto} from 'src/core';
@@ -8,17 +7,14 @@ import type {ExtensionAuto} from 'src/core';
 import {NoteNode} from './const';
 import {parserTokens} from './parser';
 import {getSchemaSpecs} from './schema';
+import type {YfmNoteSchemaOptions} from './schema';
 import {serializerTokens} from './serializer';
 
 export {NoteNode as YfmNoteNode} from './const';
 export {noteType, noteTitleType} from './utils';
+export type {YfmNoteSchemaOptions} from './schema';
 
-export type YfmNoteSpecsOptions = {
-    /**
-     * @deprecated use placeholder option in BehaviorPreset instead.
-     */
-    yfmNoteTitlePlaceholder?: NonNullable<NodeSpec['placeholder']>['content'];
-};
+export type YfmNoteSpecsOptions = YfmNoteSchemaOptions & {};
 
 export const YfmNoteSpecs: ExtensionAuto<YfmNoteSpecsOptions> = (builder, opts) => {
     const schemaSpecs = getSchemaSpecs(opts, builder.context.get('placeholder'));
