+++ packages/editor/src/markup/codemirror/react-facet.ts
@@ -1,6 +1,6 @@
 import {Facet} from '@codemirror/state';
 
-import type {ReactRenderStorage} from '../../extensions';
+import type {ReactRenderStorage} from '../../extensions/behavior/ReactRenderer';
 
 export const ReactRendererFacet = Facet.define<ReactRenderStorage, ReactRenderStorage>({
     combine: (value) => value[0],
